Moving view/learning_unit code to business/learning_unit
