+++ base/business/entity_version.py
@@ -24,6 +24,7 @@
 #
 ##############################################################################
 from base.models import entity_version, offer_year_entity
+from base.models import entity
 
 
 def find_from_offer_year(offer_year):
@@ -35,4 +36,48 @@ def find_entity_version_according_academic_year(entity_versions, academic_year):
     """This function can be use after a prefetech_related"""
     return next((entity_vers for entity_vers in entity_versions
                  if entity_vers.start_date <= academic_year.start_date and
-                    (entity_vers.end_date is None or entity_vers.end_date > academic_year.end_date)), None)
+                 (entity_vers.end_date is None or entity_vers.end_date > academic_year.end_date)), None)
+
+
+def update_entity(existing_entity, data):
+    fields_to_update = ['website', 'location', 'postal_code', 'city', 'country_id', 'phone', 'fax']
+    for f_name in fields_to_update:
+        value = data.get(f_name)
+        setattr(existing_entity, f_name, value)
+    existing_entity.save()
+
+
+def create_versions_of_existing_entity(request, same_entity):
+    new_versions_count = 0
+    entityversion_data = request.data.get('entityversion_set')
+    for version in entityversion_data:
+        identical_versions_count = entity_version.count_identical_versions(same_entity, version)
+        if not identical_versions_count:
+            parent = entity.get_by_internal_id(version.pop('parent'))
+            res = create_version(version, same_entity, parent)
+            if res is None:
+                continue
+            new_versions_count += 1
+
+    return new_versions_count
+
+
+def update_end_date_of_existing_versions(request, same_entity):
+    updated_versions_count = 0
+    entityversion_data = request.data.get('entityversion_set')
+    for version in entityversion_data:
+        to_update_versions = entity_version.find_update_candidates_versions(same_entity, version)
+        for to_update_version in to_update_versions:
+            to_update_version.end_date = version.get('end_date')
+            to_update_version.save()
+            updated_versions_count += 1
+
+    return updated_versions_count
+
+
+def create_version(version, same_entity, parent):
+    try:
+        new_version = entity_version.EntityVersion.objects.create(entity=same_entity, parent=parent, **version)
+    except AttributeError:
+        new_version = None
+    return new_version
