+++ base/models/entity_version.py
@@ -248,6 +248,8 @@ def search_entities(acronym=None, title=None, type=None):
 
 
 def find_by_id(entity_version_id):
+    if entity_version_id is None:
+        return
     return EntityVersion.objects.get(pk=entity_version_id)
