+++ base/models/learning_container_year.py
@@ -27,7 +27,7 @@ from django.contrib import admin
 from django.db import models
 from base.models import learning_unit_year
 from base.models.enums import learning_unit_year_subtypes, learning_container_year_types
-
+from reference.models import language
 
 class LearningContainerYearAdmin(admin.ModelAdmin):
     list_display = ('learning_container', 'academic_year', 'container_type', 'acronym', 'title')
@@ -61,4 +61,3 @@ class LearningContainerYear(models.Model):
 
 def find_by_id(learning_container_year_id):
     return LearningContainerYear.objects.get(pk=learning_container_year_id)
-
