+++ base/tests/views/test_entity.py
@@ -33,51 +33,70 @@ from rest_framework.reverse import reverse
 from rest_framework.test import APIClient
 from rest_framework.test import APITestCase
 
+from base.views.entity import post_entities
+
 
 class EntityViewTestCase(APITestCase):
-
     def setUp(self):
         user = UserFactory()
         token = Token.objects.create(user=user)
         self.client = APIClient()
         self.client.credentials(HTTP_AUTHORIZATION='Token ' + token.key)
 
+        self.preexisting_entity_version = EntityVersionFactory(external_id='old_external_id')
+        self.preexisting_entity_version.save()
+
     def test_create_valid_entity(self):
-        entity_version = EntityVersionFactory.build()
-        valid_entity = {
-            'organization': entity_version.entity.organization.id,
-            'external_id': entity_version.entity.external_id,
-            'entityversion_set': [
-                {
-                    "title": entity_version.title,
-                    "acronym": entity_version.acronym,
-                    "entity_type": entity_version.entity_type,
-                    "parent": entity_version.parent.id,
-                    "start_date": entity_version.start_date,
-                    "end_date": entity_version.end_date
-                }
-            ],
-            'location': entity_version.entity.location,
-            'postal_code': entity_version.entity.postal_code,
-            'city': entity_version.entity.city,
-            'country': entity_version.entity.country_id,
-            'website': entity_version.entity.website
-        }
         response = self.client.post(
-            reverse('post_entities'),
-            data=json.dumps(valid_entity),
+            reverse(post_entities),
+            data=json.dumps(_create_valid_entity()),
             content_type='application/json'
         )
 
         self.assertEqual(response.status_code, status.HTTP_201_CREATED)
+
+    def test_update_valid_entity(self):
+        response = self.client.post(
+            reverse(post_entities),
+            data=json.dumps(_create_valid_entity(self.preexisting_entity_version.entity.external_id)),
+            content_type='application/json'
+        )
+
+        self.assertEqual(response.status_code, status.HTTP_200_OK)
 
     def test_create_invalid_entity(self):
         invalid_entity = {
             'organization': 'zfeinvzepn',
         }
         response = self.client.post(
-            reverse('post_entities'),
+            reverse(post_entities),
             data=json.dumps(invalid_entity),
             content_type='application/json'
         )
         self.assertEqual(response.status_code, status.HTTP_400_BAD_REQUEST)
+
+
+def _create_valid_entity(external_id=None):
+    entity_version = EntityVersionFactory.build()
+    if not external_id:
+        external_id = entity_version.entity.external_id
+
+    return {
+        'organization': entity_version.entity.organization.id,
+        'external_id': external_id,
+        'entityversion_set': [
+            {
+                "title": entity_version.title,
+                "acronym": entity_version.acronym,
+                "entity_type": entity_version.entity_type,
+                "parent": entity_version.parent.id,
+                "start_date": entity_version.start_date,
+                "end_date": entity_version.end_date
+            }
+        ],
+        'location': entity_version.entity.location,
+        'postal_code': entity_version.entity.postal_code,
+        'city': entity_version.entity.city,
+        'country': entity_version.entity.country_id,
+        'website': entity_version.entity.website
+    }
