+++ base/urls.py
@@ -71,7 +71,7 @@ urlpatterns = [
 
     url(r'^learning_units/', include([
         url(r'^$', learning_unit.learning_units, name='learning_units'),
-        url(r'^by_activity/', learning_unit.learning_units_activity, name='learning_units_activity'),
+        url(r'^by_activity/', learning_unit.learning_units, name='learning_units_activity'),
         url(r'^by_service_course/', learning_unit.learning_units_service_course, name='learning_units_service_course'),
         url(r'^new/', include([
             url(r'^academic_year_id=(?P<academic_year>[0-9]+)$', learning_unit.learning_unit_create,
